// src/types/product.types.ts
// types/product.ts
export interface ProductType {
	id: number;
	title: string;
	description: string;
	category: string;
	price: number;
	discountPercentage: number;
	rating: number;
	stock: number;
	tags: string[];
	brand: string;
	sku: string;
	weight: number;
	dimensions: Dimensions;
	warrantyInformation: string;
	shippingInformation: string;
	availabilityStatus: string;
	reviews: Review[];
	returnPolicy: string;
	minimumOrderQuantity: number;
	meta: Meta;
	images: string[];
	thumbnail: string;
}

export interface Dimensions {
	width: number;
	height: number;
	depth: number;
}

export interface Review {
	rating: number;
	comment: string;
	date: string; // or Date if you plan to convert it
	reviewerName: string;
	reviewerEmail: string;
}

export interface Meta {
	createdAt: string; // ISO date string
	updatedAt: string;
	barcode: string;
	qrCode: string;
}